requirements.txt
pandas
numpy
requests
scikit-learn
scipy
matplotlib
seaborn
streamlit
spotipy
python-dotenv
